requirements.txt
# Core ML Libraries
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2

# Data Processing & Feature Engineering
category_encoders==2.6.3
imbalanced-learn==0.11.0
feature_engine==1.6.2

# Model Interpretation & Monitoring
shap==0.43.0
lime==0.2.0.1
evidently==0.4.14

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Web Application
streamlit==1.28.2
plotly==5.17.0
altair==5.2.0

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# MLOps & Experiment Tracking
mlflow==2.8.1
prefect==2.14.11
great-expectations==0.18.8

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Report Generation
fpdf2==2.7.6
reportlab==4.0.7
weasyprint==60.2

# Configuration & Logging
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Testing & Code Quality
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.0.0

# Utilities
requests==2.31.0
tqdm==4.66.1
click==8.1.7
python-dateutil==2.8.2

# Development
jupyter==1.0.0
ipykernel==6.26.0 